requirements.txt
Flask==0.12.2

cf-deployment-tracker==1.0.4

python-swiftclient

pyDES


python-keystoneclient